Codegen: adding a generator for verification code
To avoid naming conflicts, all global variables in each file and all
functions in each file (other than the entry point) are marked as
`static`, that way the're only visible in that file. See:

https://stackoverflow.com/questions/558122/what-is-a-static-function-in-c

Addresses #61

+++ tests/alpha.codegen.tests/xtend-gen/alpha/codegen/tests/writeC/SystemConverterManualTest.java
@@ -42,7 +42,7 @@ public class SystemConverterManualTest {
           final String alphaFile = (((("resources/" + program) + "/") + program) + ".alpha");
           final AlphaRoot root = AlphaLoader.loadAlpha(alphaFile);
           final AlphaSystem system = Utility.GetSystem(root, program);
-          final Program cAST = SystemConverter.convert(system, BaseDataType.LONG, true);
+          final Program cAST = SystemConverter.convert(system, BaseDataType.FLOAT, true);
           final String cProgram = ProgramPrinter.print(cAST).toString();
           final String cFile = (((("resources/" + program) + "/") + program) + ".c");
           final FileWriter writer = new FileWriter(cFile);
